Following Fixes have been made to PR: Still including fs-extra in package.json << fs-extra has been removed
You added the module mkdirs in package.json, but mkdirs is not an exact replacement of ensure dir. Did you mix up mkdirs the npm module and mkdirs as it is defined within fs-extra? https://github.com/jprichardson/node-fs-extra/blob/1cde1961f2e971cb42060384312c5faa419c9792/lib/mkdirs/mkdirs.js <<fs.stats w/ appropriate err statments - is equivalent to ensure dir

Will need to create a "ensureDir" function that checks if the dir is there, and if not creates it. << Tobiah having no Assets folder, is a check to make sure the person is in an actual Unity Project. Assets directory is added by default. I can create a new command if you want to Create Assets, but it should not be a part of Orginzation.

On line 84 of extension.ts, should include a question in popup asking if user would like to open the current folder instead of looking for file. <<

Line 98 of extension.ts should explain that it is looking for an asset folder in the workspace root, and to either a. open the project in the root folder or maybe b. run anyway if in the asset folder (with maybe a confirmation if they're sure?) not sure best UI here, but error messages are not clear << Clarified the Error Messages

Do a "format code" for extension.ts << formated

mkdirs is called "fs", which is confusing since it isn't fs but an extension called mkdirs. << removed.

-Kat Harris